recommend logic v0.0

+++ app/controllers/likes_controller.rb
@@ -4,7 +4,9 @@ class LikesController < ApplicationController
   
   def unlike
     like = Like.find(params[:id])
-    like.destroy
+    if like
+      like.destroy
+    end
     redirect_to :back
   end
 end
